Extract HiveWriter creation into HiveWriterFactory
